Attempt fix fastresume saving issue.

+++ src/interface.py
@@ -833,7 +833,8 @@ window, please enter your password"))
                 torrent_state = self.manager.get_torrent_state(uid)
                 if torrent_state["is_paused"]:
                     try:
-                        os.remove(self.manager.unique_IDs[uid].filename + ".fastresume")
+                        if torrent_state["state"] != 6 or torrent_state["state"] != 7:
+                            os.remove(self.manager.unique_IDs[uid].filename + ".fastresume")
                     except:
                         pass
                     self.manager.set_user_pause(uid, False, True)
@@ -849,8 +850,6 @@ window, please enter your password"))
             for uid in unique_ids:
                 self.manager.set_user_pause(uid, True, True)
                 torrent_state = self.manager.get_torrent_state(uid)
-                if torrent_state["is_paused"] == 0:
-                    self.manager.save_fastresume_data(uid)
             self.update()
         except:
             pass
